Work around duplicate key value when performing migrate-rg iid.
I got this message when trying to update the iid's:

    ERROR:  duplicate key value violates unique constraint "index_issues_on_project_id_and_iid"
    DETAIL:  Key (project_id, iid)=(6, 1770) already exists.

I worked around it by adding a constant to all existing iids before updating
them based on the issue descriptions.

The constant (6000) is arbitrary, so this PR is not really ready to be merged.
I just share it for those running into the same problem.